Preliminary version of a parent mixin class for `AnalogSignal` and `IrregularlySampledSignal` -- Issue #381
